requirements.txt
absl-py==2.2.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
aiosqlite==0.20.0
alembic==1.14.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
apache-airflow==2.10.5
apache-airflow-providers-celery==3.10.0
apache-airflow-providers-common-compat==1.3.0
apache-airflow-providers-common-io==1.5.0
apache-airflow-providers-common-sql==1.21.0
apache-airflow-providers-fab==1.5.2
apache-airflow-providers-ftp==3.12.0
apache-airflow-providers-http==5.0.0
apache-airflow-providers-imap==3.8.0
apache-airflow-providers-smtp==1.9.0
apache-airflow-providers-sqlite==4.0.0
apache-beam==2.48.0
apache-flink==1.20.0
apache-flink-libraries==1.20.0
apispec==6.8.1
argcomplete==3.5.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
async-timeout==4.0.3
attrs==25.3.0
avro-python3==1.10.2
babel==2.17.0
beautifulsoup4==4.13.3
billiard==4.2.1
black==24.10.0
bleach==6.2.0
blinker==1.9.0
cachelib==0.9.0
cachetools==5.5.2
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
clickclick==20.10.2
cloudpickle==2.2.1
colorama==0.4.6
colorlog==6.9.0
comm==0.2.2
ConfigUpdater==3.2
connexion==2.14.2
contourpy==1.3.1
crcmod==1.7
cron-descriptor==1.4.5
croniter==6.0.0
cryptography==42.0.8
cycler==0.12.1
databricks-sdk==0.50.0
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.3.1.1
distlib==0.3.9
distro==1.9.0
dj-rest-auth==6.0.0
Django==5.1.2
django-allauth==65.1.0
django-cors-headers==4.5.0
django-health-check==3.18.3
django-stubs==5.1.0
django-stubs-ext==5.1.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.7.0
docker==7.1.0
docopt==0.6.2
elastic-transport==8.17.1
elasticsearch==8.18.1
email_validator==2.2.0
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.12
fastavro==1.10.0
fasteners==0.19
fastjsonschema==2.21.1
feedparser==6.0.11
filelock==3.18.0
find_libpython==0.4.0
findspark==2.0.1
Flask==2.2.5
Flask-AppBuilder==4.5.2
Flask-Babel==2.0.0
Flask-Caching==2.3.0
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.10.1
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.2
flatbuffers==25.2.10
flower==2.0.1
fonttools==4.57.0
fqdn==1.5.1
frozenlist==1.6.0
fsspec==2025.2.0
future==1.0.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.39.0
google-auth-oauthlib==1.2.2
google-pasta==0.2.0
google-re2==1.1.20240702
googleapis-common-protos==1.66.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.1
grpcio==1.65.5
gunicorn==23.0.0
h11==0.16.0
h5py==3.13.0
hdfs==2.7.3
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
humanize==4.11.0
hyperopt==0.2.7
identify==2.6.10
idna==3.10
importlib-metadata==6.11.0
inflection==0.5.1
ipykernel==6.29.5
ipython==8.36.0
ipywidgets==8.1.6
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.9.0
jmespath==0.10.0
joblib==1.4.2
json5==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.14
kafka-python==2.0.2
keras==2.15.0
kiwisolver==1.4.8
kombu==5.4.2
langchain==0.3.25
langchain-core==0.3.61
langchain-openai==0.3.17
langchain-text-splitters==0.3.8
langsmith==0.1.147
lazy-object-proxy==1.10.0
libclang==18.1.1
limits==4.0.1
linkify-it-py==2.0.3
lockfile==0.12.2
Mako==1.3.8
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.0
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.28.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
methodtools==0.4.7
mistune==3.1.3
ml-dtypes==0.2.0
mlflow==2.21.3
mlflow-skinny==2.21.3
more-itertools==10.6.0
multidict==6.4.3
mypy_extensions==1.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nodeenv==1.9.1
notebook==7.4.1
notebook_shim==0.2.4
numpy==1.24.4
oauthlib==3.2.2
objsize==0.6.1
ollama==0.4.8
openai==1.82.0
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opt_einsum==3.4.0
ordered-set==4.1.0
orjson==3.10.16
overrides==7.7.0
packaging==24.2
pandas==2.2.0
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pemja==0.4.1
pendulum==3.0.0
pexpect==4.9.0
pgvector==0.3.6
pillow==11.2.1
platformdirs==4.3.7
pluggy==1.5.0
pre_commit==4.0.1
prison==0.2.1
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.1
proto-plus==1.26.1
protobuf==4.23.4
psutil==6.1.1
psycopg==3.2.6
psycopg-binary==3.2.6
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
py4j==0.10.9.7
pyarrow==11.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pydot==1.4.2
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.12.0
pyparsing==3.2.3
pyspark==3.5.4
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.3.0
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.4.0
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rich-argparse==1.6.0
rpds-py==0.22.3
rsa==4.9.1
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
scikit-learn==1.3.2
scipy==1.15.2
Send2Trash==1.8.3
setproctitle==1.3.4
sgmllib3k==1.0.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==1.4.49
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.46.2
tabulate==0.9.0
tenacity==9.1.2
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
terminado==0.18.1
text-unidecode==1.3
threadpoolctl==3.6.0
tiktoken==0.8.0
time-machine==2.16.0
tinycss2==1.4.0
tomli==2.2.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
types-PyYAML==6.0.12.20250402
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.1
uc-micro-py==1.0.3
universal_pathlib==0.2.6
uri-template==1.3.0
urllib3==2.4.0
uvicorn==0.34.2
vine==5.1.0
virtualenv==20.30.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==2.2.3
widgetsnbextension==4.0.14
wirerope==1.0.0
wrapt==1.14.1
WTForms==3.2.1
yarl==1.20.0
zipp==3.21.0
zstandard==0.23.0
